4.21

+++ osa4/controllers/blogs.js
@@ -59,8 +59,25 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+
+  if ( !token ) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if ( !decodedToken.id ) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if ( !blog || blog.user.toString() !== decodedToken.id.toString() ) {
+    return response.status(401).json({ error: 'you are trying to delete a non-existing blog or a blog that you are not the author to' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  return response.status(204).end()
 })
 
 module.exports = blogsRouter
